adding dev test changes

+++ lib/repository/admin_repository.dart
@@ -24,20 +24,31 @@ Future<List<Map<String, dynamic>>> fetchClassData() async {
     return [];
   }
 }
-
 Future<void> addTeacher(var attendanceData, BuildContext context) async {
+  String adminEmail = "admin@gmail.com"; // Your admin email
+  String adminPassword = "123456"; // Your admin password
+
   try {
     String email = attendanceData['email'];
-    String defaultPassword =
-        "Password123"; // Replace with your default password
+    String defaultPassword = "Password123"; 
 
     await FirebaseAuth.instance.createUserWithEmailAndPassword(
       email: email,
       password: defaultPassword,
     );
 
-    // Add the document to the 'class_attendance' collection
+    // Add teacher to Firestore
     await FirebaseFirestore.instance.collection('teachers').add(attendanceData);
+
+    // Add role
+    await FirebaseFirestore.instance
+        .collection('user_role')
+        .doc(attendanceData["email"])
+        .set(
+      {"email": attendanceData["email"], "role": "teacher"},
+      SetOptions(merge: true),
+    );
+
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: const Text(
         'Teacher added successfully! Teachers default password : Password123 Share it!',
@@ -46,6 +57,14 @@ Future<void> addTeacher(var attendanceData, BuildContext context) async {
       ),
       backgroundColor: Colors.green[400],
     ));
+
+   
+    await FirebaseAuth.instance.signOut(); // sign out the teacher user
+    await FirebaseAuth.instance.signInWithEmailAndPassword(
+      email: adminEmail,
+      password: adminPassword,
+    );
+
   } catch (e) {
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Text(
@@ -66,25 +85,63 @@ Future<void> addTeacher(var attendanceData, BuildContext context) async {
   }
 }
 
+
+
 Future<void> deleteStudent(String studentId, BuildContext context) async {
   try {
+    // Fetch the student document(s) to delete
     final querySnapshot = await FirebaseFirestore.instance
         .collection('students')
         .where('student_id', isEqualTo: studentId)
         .get();
 
-    for (var doc in querySnapshot.docs) {
-      await doc.reference.delete();
-    }
+    if (querySnapshot.docs.isNotEmpty) {
+      for (var doc in querySnapshot.docs) {
+        // Remove the student from the 'students' collection
+        await doc.reference.delete();
 
-    ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-      content: const Text(
-        'Student deleted successfully!',
-        style: TextStyle(
-            fontSize: 14, fontFamily: 'Poppins', fontWeight: FontWeight.bold),
-      ),
-      backgroundColor: Colors.green[400],
-    ));
+        // Remove the student from the 'user_role' collection
+        final studentEmail = doc.data()['email'];
+        await FirebaseFirestore.instance
+            .collection('user_role')
+            .doc(studentEmail)
+            .delete();
+
+        // Remove the student from the 'students' array in the 'classes' collection
+        final studentName = doc.data()["student_name"];
+        await FirebaseFirestore.instance
+            .collection('classes')
+            .where('students', arrayContains: {"student_id": studentId})
+            .get()
+            .then((classQuerySnapshot) async {
+          for (var classDoc in classQuerySnapshot.docs) {
+            await classDoc.reference.update({
+              "students": FieldValue.arrayRemove([
+                {"student_name": studentName, "student_id": studentId}
+              ])
+            });
+          }
+        });
+      }
+
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+        content: const Text(
+          'Student deleted successfully!',
+          style: TextStyle(
+              fontSize: 14, fontFamily: 'Poppins', fontWeight: FontWeight.bold),
+        ),
+        backgroundColor: Colors.green[400],
+      ));
+    } else {
+      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+        content: const Text(
+          'No student found with the provided ID.',
+          style: TextStyle(
+              fontSize: 14, fontFamily: 'Poppins', fontWeight: FontWeight.bold),
+        ),
+        backgroundColor: Colors.orange[400],
+      ));
+    }
   } catch (e) {
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: Text(
@@ -94,8 +151,8 @@ Future<void> deleteStudent(String studentId, BuildContext context) async {
       ),
       backgroundColor: Colors.red[400],
     ));
-  }finally{
-      Navigator.pushAndRemoveUntil(
+  } finally {
+    Navigator.pushAndRemoveUntil(
       context,
       MaterialPageRoute(
         builder: (context) => const AdminScreen(),
@@ -105,7 +162,9 @@ Future<void> deleteStudent(String studentId, BuildContext context) async {
   }
 }
 
-Future<void> updateStudentData(String studentId, Map<String, dynamic> updatedData, BuildContext context) async {
+
+Future<void> updateStudentData(String studentId,
+    Map<String, dynamic> updatedData, BuildContext context) async {
   try {
     final querySnapshot = await FirebaseFirestore.instance
         .collection('students')
@@ -154,8 +213,8 @@ Future<void> updateStudentData(String studentId, Map<String, dynamic> updatedDat
     );
   }
 }
-
-Future<void> addStudent(Map<String, dynamic> studentData, BuildContext context, String subjectId ,String classId) async {
+Future<void> addStudent(Map<String, dynamic> studentData, BuildContext context,
+    String classId) async {
   try {
     String email = studentData['email'];
     String defaultPassword = "Student123"; // Replace with your default password
@@ -166,10 +225,34 @@ Future<void> addStudent(Map<String, dynamic> studentData, BuildContext context, 
     );
 
     // Add the document to the 'students' collection
-    await FirebaseFirestore.instance.collection('students').add(studentData);
+    DocumentReference studentDocRef =
+        await FirebaseFirestore.instance.collection('students').add(studentData);
 
+    // Add the student role to the 'user_role' collection
+    await FirebaseFirestore.instance
+        .collection('user_role')
+        .doc(studentData["email"])
+        .set(
+      {"email": studentData["email"], "role": "student"},
+      SetOptions(merge: true),
+    );
 
+    // Update the class document with the student data
+    final querySnapshot = await FirebaseFirestore.instance
+        .collection('classes')
+        .where('class_id', isEqualTo: classId)
+        .get();
 
+    for (var doc in querySnapshot.docs) {
+      await doc.reference.update({
+        "students": FieldValue.arrayUnion([
+          {
+            "student_name": studentData["student_name"],
+            "student_id":studentData["student_id"],
+          }
+        ])
+      });
+    }
 
     ScaffoldMessenger.of(context).showSnackBar(SnackBar(
       content: const Text(
